General rearchitecting of the SharedData object.
Generally, the SharedData object is responsible for defining the problem dimensions and parameters. Additionally, it stores a number of gpuarrays and ndarrays, or, the data used to solve the RIME on the GPU and CPU respectively.

Since GPU code is not easy to modify or abstract, we need the ability to flexibly support different arrays for different problems, instead of blindly hardcoding all the arrays for each problem into the SharedData object. Such an approach is wasteful in terms of GPU memory.

To this end, we use Python's ability to dynamically modify object instances at runtime. Users of the SharedData object are now presented with the ability to register CPU and GPU arrays on the SharedData object. This dynamically adds the arrays, as well as any methods and members required to interact with them. Other previously hardcoded SharedData members, such as beam width and reference wavelength, are registered as properties on SharedData objects.

- BaseSharedData and GPUSharedData class collapsed into BaseSharedData.
- ndarray and gpuarray objects are created by the BaseSharedData.register_array method.
- arrays can optionally create 'uvw_shape' and 'uvw_dtype' members during registration.
- properties such "beam width", "reference wavelength" are created by the BaseSharedData.register_property method.
- Information about array shapes, types etc. and property types, defaults etc. are now printed.
- A BiroSharedData object has been created which registers arrays and properties relevent to the BIRO pipeline.
- Code for computing CPU solutions for testing has been moved into RimeCPU.py
- CPU arrays have been renamed from 'uvw' to 'uvw_cpu' for example.