Add source range to source type tuple mapping.
Given a range of sources, e.g. 30-100, the source_range
function returned the counts for each source type in
the range (and 0 for those that weren't).

Added a source_range_tuple function, the returns
the source type ranges, given a range of sources.
The source_range function has been redefined in
terms of source_range_tuple.