Reintroduce array indexing by total dimension size
The GPU kernels have two concepts of solver
dimensionality -- a maximum dimension magnitude,
hardcoded into the kernel by C/C++ defines.
In the following, the total number of timesteps
that the solver can handle is specified by

#define NTIME ${time}

within the CUDA kernel. However, we may want
to only solve a sub-portion of the problem, and
the dimension size defining this part of the problem
is defined in CUDA constant data, uploaded to the
kernel prior to execution. This is usually accessed
by the following variable

C.ntime

for example. Generally, the first case should be
used for calculating dimension strides, while the
second should be used for determining the range of
the problem to iterate over.

Previous changes went with either the first or
second approach. This change reintroduces the
distinction between the two.