Handle future completion edge case.
It may be the case that futures are removed
from queues after the queues are considered to
be too large to submit work too, but before
a decision to wait on the futures remaining in
the queue is made.

Catch any IndexErrors resulting from trying
to access empty queues and try submit the work
again.