Correct the default source keyword argument.
Changed DATA_SOURCE_DEFAULTS to DATA_SOURCE_DEFAULT and
it's value from 'defaults' to 'default'.

This was originally working because a string rather than
a kwarg was used as a key for dictionary values.
While it never made a difference on MS_test.py for v2 (all initialised from MS),
it failed for v4 (zero E and G terms).