Use copied constants instead of hardcoded values.
Stop hardcoding constants into the CUDA kernels. Instead, use constants
provided in constant memory which is copied into the kernel. Side-effect
of this change is that kernels won't need to constantly be compiled to
handle new problem sizes. However, BLOCKDIMs are still hardcoded.

Depend on hypercube 0.2.0-alpha1, which caters for arrays with
dependent sizes. We need this for npolchan=nchan*npol at the
moment.