Accumulate visibities over sources batches CPU.
Now that the GPU kernel supports source batching when
accumulating visibilities, rework the CPU code to support
the same. Previously, GPU transfers and kernels were submitted
in breadth first order to allow a CUDA stream synchronise within
the submitting thread. Now, we defer the synchronise event to
a thread specially purposed for this task. This allows us to
now enqueue our GPU transfers and kernels in depth first order.