Allow supplying + ignoring arrays during creation
In the RIMESolver.create_arrays method, add 'ignore' and
'supplied' arguments that allow the user to ignore creation
of certain arrays and supply array objects respectively.
This allows users to provide their own arrays for montblanc
to input/output from/to.

Note that supplied arrays will not be initialised, this
is the responsibility of the user.