Propagate problem chunks via descriptors.
A parameter thread now generates descriptors, which describe
the portion of the RIME currently being calculated. This descriptor
is used to update hypercubes in the feed and compute threads, which
are in turn passed to the user-defined lambda functions which supply
data to the compute functions.

The descriptor therefore allows the position in the compute space
to be identified at any point in the pipeline.