Feed parallactic angles
Cache timesteps and antenna positions on the MSRimeDataFeeder.
Would be nice to have some multi-level caching scheme for these
columns at some point, but we'd need to create new feed contexts
in the parallactic_angles method.

Creating new feed contexts basically works, but the array definitions
need to exist (and are perhaps superfluous at this point)