Handle dimension updates from source providers
Source providers can now ask the solver to change dimension sizes
at each iteration of the solver, via the
SourceProvider.updated_dimensions() hook.

The problem and memory subdivision code has been modified to
handle this at each iteration of the solver. Subdivision and
rebudgeting is avoided if less or the same memory is required.