Handle negative flux in the brightness matrix
When calculating the square root of the brightness matrix,
cater for negative flux by changing the s and t variables to
be complex valued. This should also handles cases where the determinant
is negative. However, this should not technically be possible for actual
radiation since I**2 >= Q**2 + U**2 + V**2 should hold