Distinguish between feed once and feed many data
Partition data sources into two groups, those containing
the dimensions over which we are iterating (feed many)
and those we are not (feed once).

By feeding data that is unchanged for a single computation
of the RIME, we avoid:

- Querying source providers multiple times.
- Enqueueing data multiple times.
- (Possibly) transferring data to the GPU multiple times.

This is especially beneficial in the case of the
holography beam cubes.