Supply shape functions to each Operation
Tensorflow shape functions allow tensorflow to check
the shapes of input tensors and infer the shape of output
tensors. This change provides them for each Operation,
moving duplicated tensor dimension checking code out of
the CPU and GPU ops and into the shape functions.
By inferring the shape of the Operation output, tensorflow
is able to budget memory more effectively.