Submit compute and consume futures when feeding
Tensorflow's FIFOQueues block on dequeue operations and reopening them
after they have been closed is not currently supported. Thus, to support
multiple compute threads without handling blocking,
we need a way of performing compute and consuming the result of that
compute whereby computes and consumes are exactly matched with input
queue feeding.

This commit accomplishes this by launching compute and consume
futures in their associated executors from the feed method.
This is done when a descriptor defining the input data dimensions
is received, but before the data is actually fed.

While the compute and consume futures may not end up being
exactly matched with the chunk that fed them, the number of
these operations will be correct.