Assign compute to empty shards/input queues
When the feed thread receives a descriptor, it also queries
the input queue sizes to choose the least full queue.
Assumption: this queue is associated with the least busy shard.
This queue is then fed, and compute + consumes futures are issued
that will perform work on the compute shard, as well as consume
the output of that work.