All data source signatures are now f(self,context)
Updated data source signatures to the following
- lambda c: ... changed to lambda s, c: ...
- f(ctx) changed to f(self, context)

These functions will be grafted onto the DefaultsSourceProvider which
expects them to have a signature like the above.

Also optimised many of the default/test functions to use a

```python
A = np.empty(context.shape, context.dtype)
A[:,:,:] = [[[1,0,0,1]]]
```

instead of tiling `[1,0,0,1]` up to the array size.