Separate staging areas for general feed + compute
We maintain similar staging areas on the CPU and the GPU (compute).
The CPU staging areas serve as collection points for locally and
remotely fed data, while the GPU (compute) staging areas serve as
temporary caches holding data for compute.

Functionality is not quite complete as we only the compute staging areas
are currently used.