Hacky addition of source staging area keys
We feed source chunks into a MapStagingArea, keying each chunk with an
integer key. To retrieve these chunks from the Staging Area, we need to
pass an array of these keys into tensorflow, per source type.

This commit hacks these arrays into the feed_many Staging Area, as well
as into the place where feed_many is fed with data.

Ideally these key arrays should be defined on the hypercube. They are
not for the following reasons:

1. They technically have shape something like (nsrc/chunksize,),
   which is not a hypercube dimension.

2. Technically get fed per visibility chunk, so
   *technically* they should have shape
   (nsrc/chunksize, ntime, nbl, nchan).
   This would involve instantiating an array for the entire space of
   values, which might be possible (virtually) with dask, but still.

3. The partitioning scheme currently assumes that anything with a source
   dimension belongs to that source, rather than the feed_many array.

Something should be done about this...