Rework QueueDataset
Placing the Queue in the Dataset didn't work because
_as_variant_tensor would create a new Dataset for both
the iterator creation and enqueueing ops. As no common
ancestor existed, no data would be available for the
iterator to consume.

Changed the workflow to create a QueueResource which,
in turn, is consumed by the QueueDataset, then consumed
by the Dataset iterator. Thus, QueueResource is the
common ancestor for both the Dataset and the
enqueuing operation.