Use MapStagingArea for sources instead of Datasets
After some mucking about with Datasets and tf.while_loop's,
go back to using a bare GPU MapStagingArea for
source data (lm, stokes), for the moment.
A current limitation of tensorflow iterators is that it's not
really possible to use them in a tf.while_loop construct.
So one can't really loop over a Dataset containing source info
at present.

However, do use tensorflow Datasets for the data for each visibility
chunk, which is much larger and leverages Dataset prefetch to GPU
capabilities.