tensorflow function analysis code using mock
This function allows the ability to supply a function expressing a
particular version of the RIME, constructed from the tensorflow RIME
operators.

By supplying mocked input dictionaries to the function and inspecting
queries on these dictionaries, we can infer the set of tensorflow inputs
required by the operation defined in the function.