Optionally squeeze dims and custom output schemas
In JonesMultiply, the number of output dimensions is hard-coded to 5,
so that (source,time,ant,chan,corr) can be handled.
However, the user may want some smaller output e.g. (time,corr).
Previously, dimensions missing in the input were simply assigned
a size of one in the output.

This commit allows:

1. Custom output schemas. For example (source,row,chan,corr) in
   the baseline case.
2. The ability to squeeze dimensions from the output.